Version 0.7
Former-commit-id: 1053f18691ee73b49ba0a82e1d68e50315e2e967
Former-commit-id: 81e6a645c21649e70a39f5d9a21d794ad7c32684
Former-commit-id: 1a250dabb306c7c60ebd56df04f33b4996084fa0

+++ App/Models/UsersModel.php
@@ -20,9 +20,11 @@ class UsersModel extends Model
      */
     public function create()
     {
-        $img = $this->upImg();
+        // Images are not neccessarily while registering
+        // So I'm gonna use a default avatar
+        $img = $this->defaultImg();
         // Changing image file permissions
-        chmod($this->app->file->toAvatar($img), 0777);
+//        chmod($this->app->file->toAvatar($img), 0777);
         $this->db
                 ->data('name', trim($this->request->post('name')))
                 ->data('email', trim($this->request->post('email')))
@@ -92,11 +94,23 @@ class UsersModel extends Model
     }
 
     /**
+     * Default image
+     *
+     * @return string
+     */
+    private function defaultImg()
+    {
+        $photo = rand(1, 9) . '.jpg';
+        $path  = 'default/' . $photo;
+        return $path;
+    }
+
+    /**
      * Upload image
      *
      * @return string
      */
-    public function upImg()
+    private function upImg()
     {
         $img = $this->request->file('img');
         if (!$img->exists()) {
@@ -114,9 +128,11 @@ class UsersModel extends Model
      */
     public function delete($id)
     {
-        $user    = $this->get($id);
-        $imgPath = $this->app->file->toAvatar($user->img);
-        unlink($imgPath);
+        $user = $this->get($id);
+        if (!strstr($user->img, 'default/')) {
+            $imgPath = $this->app->file->toAvatar($user->img);
+            unlink($imgPath);
+        }
         parent::delete($id);
     }
